Bugfix : Fixed some compiler warnings and some C99 and C++11 specific syntax. ( merged from GitHub, thanks to Riku Palomäki ).
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@1225 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/IFCUtil.cpp
@@ -63,7 +63,7 @@ void TempOpening::Transform(const IfcMatrix4& mat)
 // ------------------------------------------------------------------------------------------------
 aiMesh* TempMesh::ToMesh() 
 {
-	ai_assert(verts.size() == std::accumulate(vertcnt.begin(),vertcnt.end(),0));
+	ai_assert(verts.size() == std::accumulate(vertcnt.begin(),vertcnt.end(),size_t(0)));
 
 	if (verts.empty()) {
 		return NULL;
